Move model package
