+++ app/src/main/java/drewhamilton/skylight/sample/location/Location.kt
@@ -2,7 +2,7 @@ package drewhamilton.skylight.sample.location
 
 import android.content.Context
 import androidx.annotation.StringRes
-import drewhamilton.skylight.models.Coordinates
+import drewhamilton.skylight.Coordinates
 import drewhamilton.skylight.sample.R
 import java.util.*
