+++ interface/src/main/java/drewhamilton/skylight/Coordinates.kt
@@ -1,3 +1,3 @@
-package drewhamilton.skylight.models
+package drewhamilton.skylight
 
 data class Coordinates(val latitude: Double, val longitude: Double)
