+++ sso/src/main/java/drewhamilton/skylight/sso/SsoSkylight.kt
@@ -1,13 +1,13 @@
 package drewhamilton.skylight.sso
 
 import drewhamilton.skylight.Skylight
-import drewhamilton.skylight.models.AlwaysDaytime
-import drewhamilton.skylight.models.AlwaysLight
-import drewhamilton.skylight.models.Coordinates
-import drewhamilton.skylight.models.NeverDaytime
-import drewhamilton.skylight.models.NeverLight
-import drewhamilton.skylight.models.SkylightInfo
-import drewhamilton.skylight.models.Typical
+import drewhamilton.skylight.AlwaysDaytime
+import drewhamilton.skylight.AlwaysLight
+import drewhamilton.skylight.Coordinates
+import drewhamilton.skylight.NeverDaytime
+import drewhamilton.skylight.NeverLight
+import drewhamilton.skylight.SkylightInfo
+import drewhamilton.skylight.Typical
 import drewhamilton.skylight.sso.network.ApiConstants
 import drewhamilton.skylight.sso.network.InfoClient
 import drewhamilton.skylight.sso.network.models.Params
